+++ libtcc.c
@@ -388,9 +388,9 @@ ST_FUNC void section_realloc(Section *sec, unsigned long new_size)
 
 /* reserve at least 'size' bytes in section 'sec' from
    sec->data_offset. */
-ST_FUNC void *section_ptr_add(Section *sec, unsigned long size)
+ST_FUNC void *section_ptr_add(Section *sec, size_t size)
 {
-    unsigned long offset, offset1;
+    size_t offset, offset1;
 
     offset = sec->data_offset;
     offset1 = offset + size;
